+++ R/validSpeciesParams.R
@@ -49,7 +49,8 @@
 #' * `interaction_resource` is set to `1`
 #' * `n` is set to `3/4`
 #' * `p` is set to `n`
-#' 
+#' * `E_ext` is set to `0`
+#'
 #' Note that the species parameters returned by these functions are not
 #' guaranteed to produce a viable model. More checks of the parameters are
 #' performed by the individual rate-setting functions (see [setParams()] for the
@@ -71,6 +72,7 @@ validSpeciesParams <- function(species_params) {
     sp <- set_species_param_default(sp, "interaction_resource", 1)
     sp <- set_species_param_default(sp, "n", 3/4)
     sp <- set_species_param_default(sp, "p", sp$n)
+    sp <- set_species_param_default(sp, "E_ext", 0)
     return(sp)
 }